feat: add migrated files to the project; make unit tests run
