+++ Multisig/Logger/CrashlyticsProtocol.swift
