+++ Multisig/Logger/LogFormatter.swift
