+++ MultisigTests/TestUtils.swift
@@ -1,0 +1,16 @@
+//
+//  TestUtils.swift
+//  MultisigTests
+//
+//  Created by Andrey Scherbovich on 17.04.20.
+//  Copyright © 2020 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+
+/// Error that can be used in tests
+///
+/// - error: just an error
+enum TestError: Error {
+    case error
+}
